Split getCookie into server and client versions
We're gonna remove the global $c, so need to pass the request object
directly to getCookie (now getRequestCookie) on the server.
